requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flask_wtf==1.2.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
more-itertools==10.2.0
numpy==1.26.4
pandas==2.2.1
pyarrow==15.0.0
pycparser==2.21
pycryptodome==3.20.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
signxml==3.2.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
WTForms==3.1.2
